+++ _test_common/lib/in_memory_reader.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:async';
+
 import 'package:build/build.dart';
 import 'package:build_test/build_test.dart';
 
@@ -9,6 +11,15 @@ import 'package:build_runner_core/src/asset/reader.dart';
 
 class InMemoryRunnerAssetReader extends InMemoryAssetReader
     implements RunnerAssetReader {
+  final _onCanReadController = StreamController<AssetId>.broadcast();
+  Stream<AssetId> get onCanRead => _onCanReadController.stream;
+
+  @override
+  Future<bool> canRead(AssetId id) {
+    _onCanReadController.add(id);
+    return super.canRead(id);
+  }
+
   InMemoryRunnerAssetReader(
       [Map<AssetId, dynamic> sourceAssets, String rootPackage])
       : super(sourceAssets: sourceAssets, rootPackage: rootPackage);