+++ build_web_compilers/CHANGELOG.md
@@ -1,6 +1,17 @@
+## 2.15.1
+
+- Fix DDC sound mode apps to load the sound mode sdk, and only require the
+  compiled sdk dependency that it will actually use.
+
+## 2.15.0
+
+- Add some warnings for manual dart2js args that should be configured using
+  explicit options, and add back dart2js argument logging which was
+  accidentally removed.
+
 ## 2.14.0
 
-- Strip stratch directory paths from a new metadata field `fullKernelUri`
+- Strip scratch directory paths from a new metadata field `fullKernelUri`.
 
 ## 2.13.0